backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.4.0
requests==2.31.0
aiogram==3.4.1
APScheduler==3.10.4
django-redis==5.4.0
redis==5.0.1
sentry-sdk==1.40.0